pyproject.toml
[tool.poetry]
name = "habi-api"
version = "0.1.0"
description = "An api to consume Habi.co data"
authors = ["iKenshu <kenshumorales@gmail.com>"]
readme = "README.md"
packages = [{include = "habi_api"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
sqlalchemy = "^2.0.8"
mysql-connector-python = "^8.0.32"
python-dotenv = "^1.0.0"

[tool.ruff]
line-length = 79

[tool.black]
line-length = 79


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.260"
black = "^23.3.0"
pytest = "^7.2.2"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
